webapp/backend/requirements.txt
pymc>=5.0
arviz>=0.17
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
jinja2>=3.1
pyyaml>=6.0
rapidfuzz>=3.0
netcdf4>=1.6
joblib>=1.3
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.5
pydantic-settings>=2.2
google-cloud-bigquery>=3.11
google-auth>=2.28
redis>=5.0
anthropic>=0.28
httpx>=0.27
matplotlib>=3.7
torch>=2.0
